+++ opendbc/safety/modes/tesla.h
@@ -356,11 +356,11 @@ static safety_config tesla_init(uint16_t param) {
 
   SAFETY_UNUSED(param);
 #ifdef ALLOW_DEBUG
-  const int TESLA_FLAG_LONGITUDINAL_CONTROL = 1;
+  const uint16_t TESLA_FLAG_LONGITUDINAL_CONTROL = 1;
   tesla_longitudinal = GET_FLAG(param, TESLA_FLAG_LONGITUDINAL_CONTROL);
 #endif
 
-  const int TESLA_PARAM_SP_VEHICLE_BUS = 1;
+  const uint16_t TESLA_PARAM_SP_VEHICLE_BUS = 1;
 
   tesla_has_vehicle_bus = GET_FLAG(current_safety_param_sp, TESLA_PARAM_SP_VEHICLE_BUS);
